policy: Look up proxy port when creating mapstate entries
Always populate mapstate entries with the actual proxy redirection port
number, so that we can rely on a non-zero port to actually be real proxy
port number.

Note that non-realized redirects are skipped when creating mapstate
entries. When new redirects are added, the map state needs to be updated
to account for them.

Endpoint.realizedRedirects map now holds zero valued redirect ports for
Istio sidecars, so that the lookup can be made without taking endpoint's
mutex (without using Endpoint.hasSidecarProxy). Zero valued redirects are
not created or removed from the proxy package.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>