endpoint: Add proxy port to proxy stats key
Proxy stats contains the destination port of redirected traffic. When a
single port can be redirected to multiple listeners, depending on the
destination (or source), their stats entries need to be kept
separate. One way of doing this is to add the proxy port to the proxy
stats key.

Proxy port is wired through the ProxyId field in the cilium.bpf_metadata
filter config, and will be carried over to the access log messages from
there.

Proxy stats are endpoint specific, so the endpoint id need not be in
proxy stats key.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>